Move `GenotypePolyPredicate` and `GenotypeBooleanPredicate` to `gpsea.analysis.predicate.genotype` #218
