+++ src/gpsea/analysis/predicate/genotype/__init__.py
@@ -1,9 +1,13 @@
+from ._api import GenotypePolyPredicate
+# from ._api import GenotypeBooleanPredicate  # TODO: remove?
 from ._api import VariantPredicate
 from ._counter import AlleleCounter
 from ._gt_predicates import boolean_predicate, groups_predicate, recessive_predicate
 from ._variant import VariantPredicates, ProteinPredicates
 
 __all__ = [
+    'GenotypePolyPredicate',
+    # 'GenotypeBooleanPredicate',
     'boolean_predicate', 'groups_predicate', 'recessive_predicate',
     'AlleleCounter', 'VariantPredicate',
     'VariantPredicates', 'ProteinPredicates',
