+++ tests/analysis/test_mtc_filter.py
@@ -6,7 +6,8 @@ import pandas as pd
 import pytest
 
 from gpsea.analysis import HpoMtcFilter, SpecifiedTermsMtcFilter, apply_predicates_on_patients
-from gpsea.analysis.predicate import PatientCategories, GenotypePolyPredicate
+from gpsea.analysis.predicate import PatientCategories
+from gpsea.analysis.predicate.genotype import GenotypePolyPredicate
 from gpsea.analysis.predicate.phenotype import PhenotypePolyPredicate
 from gpsea.model import Cohort
